requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
google-generativeai==0.3.0
elevenlabs==0.2.26
pydantic==2.5.0
langchain-google-genai==2.1.9
langchain-core==0.3.15